+++ tests/test_render.py
@@ -373,3 +373,9 @@ class RenderFiltersTestCase(RenderTestCase):
     def test_render_base64_url_safe_decode_filter(self):
         """Test that we can render the `base64_url_safe_decode` filter correctly."""
         self._test(golden.base64_url_safe_decode_filter.cases)
+
+
+class RenderMiscTestCase(RenderTestCase):
+    def test_render_range_objects(self):
+        """Test that we can render range objects."""
+        self._test(golden.range_objects.cases)